refact: admin folder structure
M  src/app/doctoralCenter/admin/(home)/page.jsx
R  src/app/doctoralCenter/admin/_components/DoctoralCenterAdminEventManagementGrid.jsx -> src/app/doctoralCenter/admin/_components/EventManagementGrid.jsx
D  src/app/doctoralCenter/admin/_components/HomeDoctralCenterAdminGrid.jsx
R  src/app/doctoralCenter/admin/_components/DoctoralCenterAdminHomeGrid.jsx -> src/app/doctoralCenter/admin/_components/HomeGrid.jsx
R  src/app/doctoralCenter/admin/_components/DoctoralCenterAdminUnauthorizedUsersGrid.jsx -> src/app/doctoralCenter/admin/_components/UnauthorizedUsersGrid.jsx
R  src/app/doctoralCenter/admin/_components/DoctoralCenterAdminUserManagementGrid.jsx -> src/app/doctoralCenter/admin/_components/UserManagementGrid.jsx
R  src/components/config/doctoralCenter/admin/dashboard.jsx -> src/app/doctoralCenter/admin/_constants/dashboardConstants.jsx
A  src/app/doctoralCenter/admin/_constants/pathConstant.jsx
R  src/components/config/doctoralCenter/admin/sideMenuOptionsDoctoralCenterAdmin.jsx -> src/app/doctoralCenter/admin/_constants/sideMenuConstants.jsx
A  src/app/doctoralCenter/admin/_hooks/eventManagementHook.jsx
R  src/app/doctoralCenter/admin/_lib/UnauthorizedUsersGridData.jsx -> src/app/doctoralCenter/admin/_hooks/unauthorizedUsersHook.jsx
R  src/app/doctoralCenter/admin/_lib/UserManagementGridData.jsx -> src/app/doctoralCenter/admin/_hooks/userManagementHook2.jsx
D  src/app/doctoralCenter/admin/_lib/EventManagementGridData.jsx
M  src/app/doctoralCenter/admin/event-management/page.jsx
M  src/app/doctoralCenter/admin/notifications/page.jsx
M  src/app/doctoralCenter/admin/profile/page.jsx
M  src/app/doctoralCenter/admin/unauthorized-users-management/page.jsx
M  src/app/doctoralCenter/admin/user-management/page.jsx
