+++ src/app/doctoralCenter/admin/_components/HomeGrid.jsx
@@ -2,7 +2,6 @@ import Grid from "@mui/material/Grid2";
 import Box from "@mui/material/Box";
 import Stack from "@mui/material/Stack";
 import Typography from "@mui/material/Typography";
-import StatCard from "@/common/StatCard";
 import { useSelector } from "react-redux";
 import { useEffect, useState } from "react";
 import DoctoralCenterAdminAPI from "@/api/doctoralCenterAdmin";
@@ -11,24 +10,14 @@ import {
   logBarChartSeriesStruct,
   userGroupsLabelStuct,
   userGroupsPieChartStruct
-} from "@/config/doctoralCenter/admin/dashboard";
+} from "../_constants/dashboardConstants";
 import PieChartDiagram from "@/common/PieChartDiagram";
 import BarChartDashboard from "@/common/BarChartDashboard";
 import { Pagination } from "@mui/material";
 import { selectDoctoralCenter } from "@/features/user/slices/userMemoSelector";
 import Log from "@/models/Log";
 
-const statCardStruct = [
-  {
-    title: "Събития",
-    value: 0,
-    interval: "Последните 30 дни",
-    trend: "neutral",
-    data: []
-  }
-];
-
-export default function DoctoralCenterAdminHomeGrid() {
+export default function HomeGrid() {
   const doctoralCenter = useSelector(selectDoctoralCenter);
   const [userGroupsData, setUserGroupsData] = useState(userGroupsLabelStuct);
   const [userGroupsChartData, setUserGroupsChartData] = useState(
@@ -40,19 +29,18 @@ export default function DoctoralCenterAdminHomeGrid() {
   const [logs, setLogs] = useState([]);
   const [users, setUsers] = useState([]); // NOTE: Needed for the dashboard users to be rendered
   const [logsTotalSize, setLogsTotalSize] = useState(0);
-  const [statCardData, setStatCardData] = useState(statCardStruct);
   const [logsPaginaiton, setLogsPagination] = useState(1);
 
   const { fetchAutorizedUsers, fetchUnauthorizedUsers, getLogs } =
     DoctoralCenterAdminAPI();
 
   useEffect(() => {
+    let interval;
     const getUsers = async () => {
       const authUsers = await fetchAutorizedUsers();
       const unauthorizedUsers = await fetchUnauthorizedUsers();
 
       setUsers([].concat(authUsers).concat(unauthorizedUsers));
-
       setUserGroupsData(
         assignUserGroupsDataValue(
           userGroupsLabelStuct,
@@ -77,26 +65,21 @@ export default function DoctoralCenterAdminHomeGrid() {
 
     getServerLogs();
     getUsers();
-  }, [setLogs]);
+    interval = setInterval(() => {
+      getServerLogs();
+      getUsers();
+    }, process.env.NEXT_PUBLIC_FETCH_API_DURATION);
+  }, [setUsers]);
 
   useEffect(() => {
     if (logs != []) {
       setLogsChartData(assignLogsDataValue(CURRENT_YEAR));
       setSelectedYear(CURRENT_YEAR);
-      setStatCardDataset();
-
       const years = Log.getLogYears(logs);
       setLogsByYear(years);
       setLogsPagination(years.findIndex((year) => year == CURRENT_YEAR) + 1);
     }
   }, [logs]);
-
-  const setStatCardDataset = () => {
-    const statCardResult = statCardStruct;
-
-    statCardResult[0].value = logsByYear.length;
-    setStatCardData(statCardResult);
-  };
 
   const aggregateLogsByYearMonths = (logLevel, year) => {
     const levelSpecificLogs = Log.filterByLevelAndYear(logs, logLevel, year);
@@ -138,9 +121,8 @@ export default function DoctoralCenterAdminHomeGrid() {
     authUsers,
     unauthUsers
   ) => {
-    const unauthorizedCount = unauthUsers.length;
     const users = ["phd", "committee", "manager", "expert", "admin"];
-    const userCounts = [unauthorizedCount];
+    const userCounts = [unauthUsers.length];
 
     users.forEach((userGroup) => {
       const userCount = authUsers.filter(
@@ -169,15 +151,6 @@ export default function DoctoralCenterAdminHomeGrid() {
         Добре дошли {doctoralCenter.name}
       </Typography>
 
-      {Array.isArray(statCardData) ? (
-        statCardData.map((card, index) => (
-          <Grid key={index} size={{ xs: 12, sm: 6, lg: 3 }}>
-            <StatCard {...card} />
-          </Grid>
-        ))
-      ) : (
-        <></>
-      )}
       <Grid
         container
         spacing={2}
@@ -209,10 +182,6 @@ export default function DoctoralCenterAdminHomeGrid() {
             onChange={logYearChangeOnClick}
           />
         </Grid>
-      </Grid>
-
-      <Grid container spacing={2} columns={12}>
-        <Grid size={{ xs: 12, lg: 9 }}>{/*Data grid is here*/}</Grid>
       </Grid>
     </Box>
   );
