+++ src/app/doctoralCenter/admin/_constants/pathConstant.jsx
@@ -1,0 +1,1 @@
+export const path = "/doctoralCenter/admin";
