+++ src/app/doctoralCenter/admin/event-management/page.jsx
@@ -1,16 +1,18 @@
 "use client";
 import Layout from "@/components/main-layout/Layout";
 
-import { sideMenuOptionsDoctoralCenterAdmin } from "@/config/doctoralCenter/admin/sideMenuOptionsDoctoralCenterAdmin";
-import DoctoralCenterAdminEventManagementGrid from "../_components/DoctoralCenterAdminEventManagementGrid";
+import EventManagementGrid from "../_components/EventManagementGrid";
+import { sideMenu } from "../_constants/sideMenuConstants";
+
+import { path } from "../_constants/pathConstant";
 
 export default function Page() {
   return (
     <Layout
       headerTitle={"Управление на събития"}
-      basePath={"/doctoralCenter/admin"}
-      MainView={<DoctoralCenterAdminEventManagementGrid />}
-      mainListItems={sideMenuOptionsDoctoralCenterAdmin}
+      basePath={path}
+      MainView={<EventManagementGrid />}
+      mainListItems={sideMenu}
     />
   );
 }
