+++ src/app/doctoralCenter/admin/unauthorized-users-management/page.jsx
@@ -1,16 +1,17 @@
 "use client";
 import Layout from "@/components/main-layout/Layout";
 
-import { sideMenuOptionsDoctoralCenterAdmin } from "@/config/doctoralCenter/admin/sideMenuOptionsDoctoralCenterAdmin";
-import DoctoralCenterAdminUnauthorizedUsersGrid from "../_components/DoctoralCenterAdminUnauthorizedUsersGrid";
+import UnauthorizedUsersGrid from "../_components/UnauthorizedUsersGrid";
+import { sideMenu } from "../_constants/sideMenuConstants";
+import { path } from "../_constants/pathConstant";
 
 export default function Page() {
   return (
     <Layout
       headerTitle={"Неудостоверени потребители"}
-      basePath={"/doctoralCenter/admin"}
-      MainView={<DoctoralCenterAdminUnauthorizedUsersGrid />}
-      mainListItems={sideMenuOptionsDoctoralCenterAdmin}
+      basePath={path}
+      MainView={<UnauthorizedUsersGrid />}
+      mainListItems={sideMenu}
     />
   );
 }
